Integralização da estrutura de visualização de extratos de conta dentro do Extrato, descartando o uso do ListaConta e do Conta. Correção do problema do NaN dentro do Extrato. Fidex #25.
